# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Bravo API
  version: 1.0.0
  description: OpenAPI documentation for the Bravo backend.
paths:
  /api/auth/login/:
    post:
      operationId: auth_login_create
      description: Authenticates the user with email and password and returns JWT
        access and refresh tokens.
      summary: Login
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailTokenObtainPair'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/EmailTokenObtainPair'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/EmailTokenObtainPair'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailTokenObtainPair'
          description: ''
  /api/auth/me/:
    get:
      operationId: auth_me_retrieve
      description: Returns the authenticated user's data based on the token sent in
        the request.
      summary: Get authenticated user
      tags:
      - auth
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
  /api/auth/register/:
    post:
      operationId: auth_register_create
      description: Creates a new user account and returns the registered profile along
        with JWT tokens.
      summary: Register user
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Register'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Register'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Register'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Register'
          description: ''
  /api/auth/token/refresh/:
    post:
      operationId: auth_token_refresh_create
      description: Receives a valid refresh token and returns a new JWT access token.
      summary: Refresh token
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefresh'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenRefresh'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TokenRefresh'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefresh'
          description: ''
  /api/auth/verify-email/:
    post:
      operationId: auth_verify_email_create
      description: Validates the verification token and marks the user's email as
        verified.
      summary: Verify email
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmail'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/VerifyEmail'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/VerifyEmail'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyEmail'
          description: ''
...
components:
  schemas:
    AccountStatusEnum:
      enum:
      - active
      - inactive
      - suspended
      type: string
      description: |-
        * `active` - Active
        * `inactive` - Inactive
        * `suspended` - Suspended
...
    EmailTokenObtainPair:
      type: object
      properties:
        email:
          type: string
          writeOnly: true
        password:
          type: string
          writeOnly: true
      required:
      - email
      - password
...
    Register:
      type: object
      properties:
        username:
          type: string
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          writeOnly: true
          minLength: 8
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
      required:
      - email
      - password
      - username
...
    TokenRefresh:
      type: object
      properties:
        access:
          type: string
          readOnly: true
        refresh:
          type: string
      required:
      - access
      - refresh
    User:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        uuid:
          type: string
          format: uuid
          readOnly: true
        username:
          type: string
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 254
        email_verified:
          type: boolean
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        status:
          $ref: '#/components/schemas/AccountStatusEnum'
      required:
      - email
      - id
      - username
      - uuid
    VerifyEmail:
      type: object
      properties:
        token:
          type: string
      required:
      - token
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT